requirements.txt
streamlit==1.28.2
yt-dlp==2024.3.9
pytube==15.0.0
transformers==4.36.2
openai==1.12.0
python-dotenv==1.0.0
torch==2.5.1
torchaudio
pandas
ffmpeg-python
requests
numpy
